requirements.txt
# ICR Performance Test Dependencies
paramiko>=2.12.0
aiohttp>=3.8.0
psycopg2-binary>=2.9.0
openpyxl>=3.0.0
pandas>=1.5.0
pyinstaller>=5.0.0
